warnings--.  Playlistgraphicsitem is not fun to touch :/
svn path=/trunk/extragear/multimedia/amarok/; revision=788149

+++ src/context/Applet.cpp
@@ -59,7 +59,7 @@ QFont Context::Applet::shrinkTextSizeToFit( const QString& text, const QRectF& b
 QString Context::Applet::truncateTextToFit( QString text, const QFont& font, const QRectF& bounds )
 {
     QFontMetrics fm( font );
-    return fm.elidedText ( text, Qt::ElideRight, bounds.width() );
+    return fm.elidedText ( text, Qt::ElideRight, (int)bounds.width() );
     
 }
